requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cattrs==25.1.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
platformdirs==4.3.8
requests==2.32.4
requests-cache==1.2.1
soupsieve==2.7
typing==3.7.4.3
typing_extensions==4.14.0
url-normalize==2.2.1
urllib3==2.4.0
Werkzeug==3.1.3
